# openapi-v1.json
{
  "components": {
    "schemas": {
...
      "BeamModel": {
        "properties": {
          "machine_id": {
            "title": "Machine Id",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "machine_id",
          "version",
          "sha256"
        ],
        "title": "BeamModel",
        "type": "object"
      },
      "BeamModelBody": {
        "properties": {
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "parameters"
        ],
        "title": "BeamModelBody",
        "type": "object"
      },
      "Body_upload_measurement_api_v1_commissioning_upload_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_measurement_api_v1_commissioning_upload_post",
        "type": "object"
      },
...
      "CreateMachineBody": {
        "properties": {
          "engine": {
            "title": "Engine",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "params": {
            "additionalProperties": true,
            "title": "Params",
            "type": "object"
          }
        },
        "required": [
          "name",
          "engine",
          "params"
        ],
        "title": "CreateMachineBody",
        "type": "object"
      },
...
      "DoseResult": {
        "description": "Returned by ``dose_service.run_dose_calc`` and ``run_phantom_calc``.\n\nThe service layer uses camelCase keys; ``validation_alias`` lets Pydantic\naccept those keys while keeping Pythonic field names here.",
        "properties": {
          "max_dose_gy": {
            "title": "Max Dose Gy",
            "type": "number"
          },
          "mean_dose_gy": {
            "title": "Mean Dose Gy",
            "type": "number"
          },
          "n_projections": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "N Projections"
          },
          "provenance": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProvenanceManifest"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provenance"
          }
        },
        "required": [
          "max_dose_gy",
          "mean_dose_gy"
        ],
        "title": "DoseResult",
        "type": "object"
      },
...
      "GoldenData": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "machine_id": {
            "title": "Machine Id",
            "type": "string"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "machine_id"
        ],
        "title": "GoldenData",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MachineRecord": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "engine": {
            "title": "Engine",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "locked_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Locked Hash"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "params": {
            "title": "Params"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "id",
          "name",
          "engine",
          "status",
          "params",
          "created_at"
        ],
        "title": "MachineRecord",
        "type": "object"
      },
...
      "PhantomCalcBody": {
        "properties": {
          "engine": {
            "title": "Engine",
            "type": "string"
          },
          "machine_id": {
            "title": "Machine Id",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": true,
            "title": "Parameters",
            "type": "object"
          }
        },
        "required": [
          "machine_id",
          "engine",
          "parameters"
        ],
        "title": "PhantomCalcBody",
        "type": "object"
      },
      "ProvenanceManifest": {
        "properties": {
          "engine": {
            "additionalProperties": true,
            "title": "Engine",
            "type": "object"
          },
          "git_sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Git Sha"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "integer"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "sidecar_version": {
            "title": "Sidecar Version",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "sidecar_version",
          "session_id"
        ],
        "title": "ProvenanceManifest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Axioma Sidecar",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/commissioning/calculate_water_phantom": {
      "post": {
        "operationId": "calculate_water_phantom_api_v1_commissioning_calculate_water_phantom_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PhantomCalcBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DoseResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Calculate Water Phantom"
      }
    },
    "/api/v1/commissioning/lock": {
      "post": {
        "operationId": "lock_machine_api_v1_commissioning_lock_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lock Machine"
      }
    },
    "/api/v1/commissioning/machines": {
      "get": {
        "operationId": "list_machines_api_v1_commissioning_machines_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MachineRecord"
                  },
                  "title": "Response List Machines Api V1 Commissioning Machines Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Machines"
      },
      "post": {
        "operationId": "create_machine_api_v1_commissioning_machines_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMachineBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MachineRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Machine"
      }
    },
    "/api/v1/commissioning/machines/{machine_id}/beam-model": {
      "get": {
        "operationId": "get_beam_model_api_v1_commissioning_machines__machine_id__beam_model_get",
        "parameters": [
          {
            "in": "path",
            "name": "machine_id",
            "required": true,
            "schema": {
              "title": "Machine Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeamModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Beam Model"
      },
      "put": {
        "operationId": "put_beam_model_api_v1_commissioning_machines__machine_id__beam_model_put",
        "parameters": [
          {
            "in": "path",
            "name": "machine_id",
            "required": true,
            "schema": {
              "title": "Machine Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BeamModelBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeamModel"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Beam Model"
      }
    },
    "/api/v1/commissioning/machines/{machine_id}/golden-data": {
      "get": {
        "operationId": "get_golden_data_api_v1_commissioning_machines__machine_id__golden_data_get",
        "parameters": [
          {
            "in": "path",
            "name": "machine_id",
            "required": true,
            "schema": {
              "title": "Machine Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoldenData"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Golden Data"
      }
    },
    "/api/v1/commissioning/upload": {
      "post": {
        "operationId": "upload_measurement_api_v1_commissioning_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_measurement_api_v1_commissioning_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Measurement"
      }
    },
...
  }
}